recmusic/requirements.txt
Django==1.6
pymongo==2.8
numpy==1.9.2
csc-pysparse==1.1.1.4
csc-utils==0.6.7
decorator==3.4.2
networkx==1.9.1
Divisi2==2.2.5
boto==2.38.0
django-mongodb-engine==0.6.0
djangotoolbox==1.6.2
django-dbindexer==1.6.0
https://github.com/django-nonrel/django/tarball/nonrel-1.6
mongoengine==0.8.0
tweepy==3.3.0
ordereddict==1.1
importlib==1.0.3
south==1.0.2